refactor(api): extract the v1 public-API contract into @mcpjam/sdk/public-api
The error-code union, status map, internal->public mapping, and envelope
builders were hand-duplicated between the Inspector gateway
(server/routes/v1/contract.ts) and the Convex backend, kept in sync only by
byte-identical golden fixtures. Promote the inspector copy into a pure,
dependency-free SDK subpath so there is one source of truth.

- sdk/src/public-api/index.ts: the contract, framework-agnostic (no Hono /
  Convex / Node imports), with a tsup entry, a `@mcpjam/sdk/public-api` export,
  and a packaging-test assertion.
- server/routes/v1/contract.ts now re-exports the subpath; the Hono response
  adapters stay in envelope.ts. server/vitest.config.ts aliases the subpath to
  source like the other first-party subpaths.

The golden fixtures + contract.test.ts remain the cross-surface guard: that
suite now validates the re-exported SDK values against the fixtures, proving the
subpath is byte-identical to the retired inspector copy and the backend's
still-local copy.

The Convex backend keeps its local copy until its pinned @mcpjam/sdk version is
bumped to one that ships this subpath (separate follow-up, since the backend
consumes the SDK as a published dependency).
