+++ sdk/package.json
@@ -67,6 +67,11 @@
       "import": "./dist/platform/index.js",
       "default": "./dist/platform/index.js"
     },
+    "./public-api": {
+      "types": "./dist/public-api/index.d.ts",
+      "import": "./dist/public-api/index.js",
+      "default": "./dist/public-api/index.js"
+    },
     "./widget-runtime": {
       "types": "./dist/widget-runtime/index.d.ts",
       "import": "./dist/widget-runtime/index.js",
@@ -86,7 +91,7 @@
     "test": "vitest run",
     "test:watch": "vitest",
     "test:coverage": "vitest run --coverage",
-    "test:packaging": "npm run build && node --input-type=module -e \"await import('@mcpjam/sdk'); await import('@mcpjam/sdk/browser'); await import('@mcpjam/sdk/worker'); await import('@mcpjam/sdk/operations'); await import('@mcpjam/sdk/skill-reference'); await import('@mcpjam/sdk/model-factory'); await import('@mcpjam/sdk/matchers'); await import('@mcpjam/sdk/predicates'); await import('@mcpjam/sdk/host-config'); const w = await import('@mcpjam/sdk/widget-runtime'); for (const n of ['getToolVisibility','isVisibleToModelOnly','isVisibleToAppOnly','LoggingTransport','buildOuterAllowAttribute','buildOuterSandboxAttribute','resolveIframeSandboxPolicy','createHostAppBridge','registerHostBridgeHandlers']) { if (typeof w[n] !== 'function') { throw new Error('missing widget-runtime export: ' + n); } } if (w['DEFAULT_IFRAME_SANDBOX'] === undefined) { throw new Error('missing widget-runtime export: DEFAULT_IFRAME_SANDBOX'); } const p = await import('@mcpjam/sdk/platform'); for (const n of ['PlatformApiClient','PlatformApiError','buildShowServersPayload','resolveProject','listProjectsOperation','listProjectServersOperation','showServersOperation','createTunnelOperation','closeTunnelOperation']) { if (p[n] === undefined) { throw new Error('missing platform export: ' + n); } } const m = await import('@mcpjam/sdk/host-config/internal'); for (const n of ['canonicalizeHostConfigV2','computeHostConfigHashV2','isAppOnlyTool','filterAppOnlyTools','applyVisibilityPolicyAndCountSignals','extractHostExecutionPolicy','buildHostIterationMetadata','resolveOpenAiCompatForHostConfig','readOpenAiCompatOverride','compatPresetForHostStyle','normalizeSdkEvalHostConfigForWire']) { if (typeof m[n] !== 'function') { throw new Error('missing export: ' + n); } }\"",
+    "test:packaging": "npm run build && node --input-type=module -e \"await import('@mcpjam/sdk'); await import('@mcpjam/sdk/browser'); await import('@mcpjam/sdk/worker'); await import('@mcpjam/sdk/operations'); await import('@mcpjam/sdk/skill-reference'); await import('@mcpjam/sdk/model-factory'); await import('@mcpjam/sdk/matchers'); await import('@mcpjam/sdk/predicates'); await import('@mcpjam/sdk/host-config'); const w = await import('@mcpjam/sdk/widget-runtime'); for (const n of ['getToolVisibility','isVisibleToModelOnly','isVisibleToAppOnly','LoggingTransport','buildOuterAllowAttribute','buildOuterSandboxAttribute','resolveIframeSandboxPolicy','createHostAppBridge','registerHostBridgeHandlers']) { if (typeof w[n] !== 'function') { throw new Error('missing widget-runtime export: ' + n); } } if (w['DEFAULT_IFRAME_SANDBOX'] === undefined) { throw new Error('missing widget-runtime export: DEFAULT_IFRAME_SANDBOX'); } const p = await import('@mcpjam/sdk/platform'); for (const n of ['PlatformApiClient','PlatformApiError','buildShowServersPayload','resolveProject','listProjectsOperation','listProjectServersOperation','showServersOperation','createTunnelOperation','closeTunnelOperation']) { if (p[n] === undefined) { throw new Error('missing platform export: ' + n); } } const m = await import('@mcpjam/sdk/host-config/internal'); for (const n of ['canonicalizeHostConfigV2','computeHostConfigHashV2','isAppOnlyTool','filterAppOnlyTools','applyVisibilityPolicyAndCountSignals','extractHostExecutionPolicy','buildHostIterationMetadata','resolveOpenAiCompatForHostConfig','readOpenAiCompatOverride','compatPresetForHostStyle','normalizeSdkEvalHostConfigForWire']) { if (typeof m[n] !== 'function') { throw new Error('missing export: ' + n); } } const pub = await import('@mcpjam/sdk/public-api'); for (const n of ['isV1ErrorCode','mapInternalCode','v1ErrorBody','v1Page','classifyRuntimeError']) { if (typeof pub[n] !== 'function') { throw new Error('missing public-api export: ' + n); } } for (const n of ['V1_ERROR_CODES','V1_ERROR_STATUS','INTERNAL_TO_V1_CODE']) { if (pub[n] === undefined) { throw new Error('missing public-api export: ' + n); } }\"",
     "lint": "eslint src tests",
     "lint:fix": "eslint src tests --fix",
     "format": "prettier --write \"src/**/*.ts\" \"tests/**/*.ts\"",
