+++ sdk/tsup.config.ts
@@ -25,6 +25,9 @@ export default defineConfig({
     "src/host-config/templates/index.ts",
     // Runtime-agnostic Platform API client (Workers/browser/Node safe).
     "src/platform/index.ts",
+    // Framework-free public-API wire contract (error codes + envelopes) shared
+    // by the Inspector gateway and (eventually) the Convex backend.
+    "src/public-api/index.ts",
     // Framework-free widget/app runtime building blocks (SEP-1865).
     "src/widget-runtime/index.ts",
   ],
